+++ CHANGELOG.md
@@ -11,6 +11,24 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Deprecated
 #### Removed
 
+## [2.0.0] - 2025-02-03
+
+### Added
+
+* feat: Add `licenses.others` option
+* feat: Add `licenses.unknown` option
+* feat: Add `production` option
+* feat: Add `development` option
+* feat: Add `direct` option
+* feat: Add `packages` option
+* feat: Add `excludePackages` option
+* feat: Add `excludePrivatePackages` option
+
+### Removed
+
+* feat(BREAKING CHANGE): Remove `allowWarnings` option. It has no sense anymore with the new options.
+* feat(BREAKING CHANGE): Remove `licenseCheckerOptions` option. It has no sense anymore with the new options.
+
 ## [1.0.0] - 2025-01-17
 
 #### Added